bot/requirements.txt
asyncpg==0.29.0
hikari==2.0.0
hikari-arc==1.3.4
hikari-miru==4.1.1
human-readable==1.3.4
parsedatetime==2.6
